+++ core/Cargo.toml
@@ -46,14 +46,14 @@ ring = { version = "0.16.9", features = ["alloc", "std"], default-features = fal
 async-std = { version = "1.6.2", features = ["attributes"] }
 base64 = "0.13.0"
 criterion = "0.3"
-libp2p-mplex = { path = "../muxers/mplex" }
-libp2p-noise = { path = "../transports/noise" }
-libp2p-tcp = { path = "../transports/tcp" }
 serde_json = "1.0"
 rmp-serde = "1.0"
 multihash = { version = "0.16", default-features = false, features = ["arb"] }
-quickcheck = "0.9.0"
 rand07 = { package = "rand", version = "0.7" }
+quickcheck = "0.9.0"
+libp2p-mplex = { path = "../muxers/mplex" }
+libp2p-noise = { path = "../transports/noise" }
+libp2p-tcp = { path = "../transports/tcp" }
 
 [build-dependencies]
 prost-build = "0.9"
