+++ core/src/lib.rs
@@ -61,18 +61,16 @@ pub mod connection;
 pub mod either;
 pub mod identity;
 pub mod muxing;
-pub mod network;
 pub mod peer_record;
 pub mod signed_envelope;
 pub mod transport;
 pub mod upgrade;
 
-pub use connection::{Connected, ConnectedPoint, Endpoint};
+pub use connection::{ConnectedPoint, Endpoint};
 pub use identity::PublicKey;
 pub use multiaddr::Multiaddr;
 pub use multihash;
 pub use muxing::StreamMuxer;
-pub use network::{DialOpts, Network};
 pub use peer_id::PeerId;
 pub use peer_record::PeerRecord;
 pub use signed_envelope::SignedEnvelope;
