+++ misc/multistream-select/Cargo.toml
@@ -22,6 +22,7 @@ unsigned-varint = "0.7"
 async-std = "1.6.2"
 env_logger = "0.9"
 libp2p-core = { path = "../../core", default-features = false }
+libp2p-swarm = { path = "../../swarm", default-features = false }
 libp2p-mplex = { path = "../../muxers/mplex" }
 libp2p-plaintext = { path = "../../transports/plaintext" }
 quickcheck = "0.9.0"
