+++ swarm/src/connection/error.rs
